requirements.txt
-i https://pypi.org/simple
alembic==1.14.0; python_version >= '3.8'
aniso8601==9.0.1
argparse==1.4.0
bcrypt==4.2.0; python_version >= '3.7'
blinker==1.9.0; python_version >= '3.9'
certifi==2024.8.30; python_version >= '3.6'
charset-normalizer==3.4.0; python_full_version >= '3.7.0'
click==8.1.7; python_version >= '3.7'
cloudinary==1.41.0
colorama==0.4.6; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5, 3.6'
cors==1.0.1
filelock==3.16.1; python_version >= '3.8'
flask==3.1.0; python_version >= '3.9'
flask-admin==1.6.1; python_version >= '3.6'
flask-cors==5.0.0
flask-jwt-extended==4.6.0; python_version >= '3.7' and python_version < '4'
flask-login==0.6.3; python_version >= '3.7'
flask-migrate==4.0.7; python_version >= '3.6'
flask-redis==0.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
flask-restful==0.3.10
flask-serializer==0.0.5.1; python_version >= '2.7'
flask-sqlalchemy==3.1.1; python_version >= '3.8'
future==1.0.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
gevent==24.11.1; python_version >= '3.9'
greenlet==3.1.1; platform_machine == 'aarch64' or (platform_machine == 'ppc64le' or (platform_machine == 'x86_64' or (platform_machine == 'amd64' or (platform_machine == 'AMD64' or (platform_machine == 'win32' or platform_machine == 'WIN32')))))
gunicorn==23.0.0; python_version >= '3.7'
idna==3.10; python_version >= '3.6'
itsdangerous==2.2.0; python_version >= '3.8'
jinja2==3.1.4; python_version >= '3.7'
mako==1.3.6; python_version >= '3.8'
markupsafe==3.0.2; python_version >= '3.9'
marshmallow==3.23.1; python_version >= '3.9'
packaging==24.2; python_version >= '3.8'
passlib==1.7.4
psycopg2-binary==2.9.9; python_version >= '3.7'
pyjwt==2.9.0; python_version >= '3.8'
pysocks==1.7.1; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-dotenv==1.0.1; python_version >= '3.8'
pytz==2024.2
redis==5.2.0; python_version >= '3.8'
requests==2.32.3; python_version >= '3.8'
requests-file==2.1.0
setuptools==70.3.0; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlalchemy==2.0.29; python_version >= '3.7'
sqlalchemy-serializer==1.4.22; python_version >= '3.10' and python_version < '4.0'
tldextract==5.1.3; python_version >= '3.9'
typing-extensions==4.12.2; python_version >= '3.8'
urllib3==2.2.3; python_version >= '3.8'
werkzeug==3.1.3; python_version >= '3.9'
wtforms==3.2.1; python_version >= '3.9'
zope.event==5.0; python_version >= '3.7'
zope.interface==7.1.1; python_version >= '3.8'
pytest
pytest-flask
flask-testing
redis